// web/src/lib/api/generated/models/internal-customers-create-customer-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POS Kasir API
 * POS Kasir API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface InternalCustomersCreateCustomerRequest {
    'address'?: string;
    'email'?: string;
    'name': string;
    'phone'?: string;
}